Added Check According to Role Permission & Update Datagrid Modal Support As Dynamic

+++ packages/Webkul/Admin/src/DataGrids/CMS/CMSPageDataGrid.php
@@ -87,23 +87,27 @@ class CMSPageDataGrid extends DataGrid
             },
         ]);
 
-        $this->addAction([
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.cms.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => function ($row) {
-                return route('admin.cms.edit', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('cms.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.cms.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => function ($row) {
+                    return route('admin.cms.edit', $row->id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.cms.index.datagrid.delete'),
-            'method' => 'DELETE',
-            'url'    => function ($row) {
-                return route('admin.cms.delete', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('cms.delete')) {
+            $this->addAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.cms.index.datagrid.delete'),
+                'method' => 'DELETE',
+                'url'    => function ($row) {
+                    return route('admin.cms.delete', $row->id);
+                },
+            ]);
+        }
     }
 
     /**
@@ -113,10 +117,12 @@ class CMSPageDataGrid extends DataGrid
      */
     public function prepareMassActions()
     {
-        $this->addMassAction([
-            'title'  => trans('admin::app.cms.index.datagrid.delete'),
-            'method' => 'POST',
-            'url'    => route('admin.cms.mass_delete'),
-        ]);
+        if (bouncer()->hasPermission('cms.mass-delete')) {
+            $this->addMassAction([
+                'title'  => trans('admin::app.cms.index.datagrid.delete'),
+                'method' => 'POST',
+                'url'    => route('admin.cms.mass_delete'),
+            ]);
+        }
     }
 }
